feat: added path to the puppet file

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,7 @@
 # This manifest kilss a process called killmenow
 
 exec {
-  command     => 'pkill killmenow',
-  refreshonly => true,
+  command => 'pkill killmenow',
+  path  => '/usr/bin',
+  onlyif  => 'pgrep killmenow',
 }
